backend/requirements.txt
# フレームワーク関連の依存関係
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.24.1
transformers>=4.21.0
torch>=1.12.0
chromadb>=0.4.0
numpy<2.0.0,>=1.21.0
sentence-transformers>=2.2.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0 

# データベース関連の依存関係
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
email-validator>=2.0.0

# パスワードハッシュ化
passlib[bcrypt]>=1.7.4

# JWT認証
python-jose[cryptography]>=3.3.0

# テスト関連の依存関係
pytest==8.3.5
pytest-cov==4.0.0
flake8==6.0.0